+++ src/backend/sqlite/table.rs
@@ -1,8 +1,8 @@
 use super::*;
 
 impl TableBuilder for SqliteQueryBuilder {
-    fn prepare_column_def(&self, column_def: &ColumnDef, sql: &mut SqlWriter) {
-        column_def.name.prepare(sql, self.quote());
+    fn prepare_column_def(&self, column_def: &ColumnDef, sql: &mut dyn SqlWriter) {
+        column_def.name.prepare(sql.as_writer(), self.quote());
 
         if let Some(column_type) = &column_def.types {
             write!(sql, " ").unwrap();
@@ -35,7 +35,7 @@ impl TableBuilder for SqliteQueryBuilder {
         }
     }
 
-    fn prepare_column_type(&self, column_type: &ColumnType, sql: &mut SqlWriter) {
+    fn prepare_column_type(&self, column_type: &ColumnType, sql: &mut dyn SqlWriter) {
         write!(
             sql,
             "{}",
@@ -102,7 +102,7 @@ impl TableBuilder for SqliteQueryBuilder {
                     _ => "blob".into(),
                 },
                 ColumnType::VarBinary(length) => format!("binary({})", length),
-                ColumnType::Boolean => "integer".into(),
+                ColumnType::Boolean => "boolean".into(),
                 ColumnType::Money(precision) => match precision {
                     Some((precision, scale)) => format!("integer({}, {})", precision, scale),
                     None => "integer".into(),
@@ -121,24 +121,15 @@ impl TableBuilder for SqliteQueryBuilder {
         .unwrap()
     }
 
-    fn prepare_column_spec(&self, column_spec: &ColumnSpec, sql: &mut SqlWriter) {
-        match column_spec {
-            ColumnSpec::Null => write!(sql, "NULL"),
-            ColumnSpec::NotNull => write!(sql, "NOT NULL"),
-            ColumnSpec::Default(value) => write!(sql, "DEFAULT {}", self.value_to_string(value)),
-            ColumnSpec::AutoIncrement => write!(sql, "AUTOINCREMENT"),
-            ColumnSpec::UniqueKey => write!(sql, "UNIQUE"),
-            ColumnSpec::PrimaryKey => write!(sql, "PRIMARY KEY"),
-            ColumnSpec::Extra(string) => write!(sql, "{}", string),
-        }
-        .unwrap()
+    fn column_spec_auto_increment_keyword(&self) -> &str {
+        "AUTOINCREMENT"
     }
 
-    fn prepare_table_drop_opt(&self, _drop_opt: &TableDropOpt, _sql: &mut dyn std::fmt::Write) {
+    fn prepare_table_drop_opt(&self, _drop_opt: &TableDropOpt, _sql: &mut dyn SqlWriter) {
         // SQLite does not support table drop options
     }
 
-    fn prepare_table_alter_statement(&self, alter: &TableAlterStatement, sql: &mut SqlWriter) {
+    fn prepare_table_alter_statement(&self, alter: &TableAlterStatement, sql: &mut dyn SqlWriter) {
         if alter.options.is_empty() {
             panic!("No alter option found")
         };
@@ -160,9 +151,9 @@ impl TableBuilder for SqliteQueryBuilder {
             }
             TableAlterOption::RenameColumn(from_name, to_name) => {
                 write!(sql, "RENAME COLUMN ").unwrap();
-                from_name.prepare(sql, self.quote());
+                from_name.prepare(sql.as_writer(), self.quote());
                 write!(sql, " TO ").unwrap();
-                to_name.prepare(sql, self.quote());
+                to_name.prepare(sql.as_writer(), self.quote());
             }
             TableAlterOption::DropColumn(_) => {
                 panic!("Sqlite not support dropping table column")
@@ -176,7 +167,11 @@ impl TableBuilder for SqliteQueryBuilder {
         }
     }
 
-    fn prepare_table_rename_statement(&self, rename: &TableRenameStatement, sql: &mut SqlWriter) {
+    fn prepare_table_rename_statement(
+        &self,
+        rename: &TableRenameStatement,
+        sql: &mut dyn SqlWriter,
+    ) {
         write!(sql, "ALTER TABLE ").unwrap();
         if let Some(from_name) = &rename.from_name {
             self.prepare_table_ref_table_stmt(from_name, sql);