+++ src/func.rs
@@ -41,7 +41,7 @@ impl Func {
     /// struct MyFunction;
     ///
     /// impl Iden for MyFunction {
-    ///     fn unquoted(&self, s: &mut dyn FmtWrite) {
+    ///     fn unquoted(&self, s: &mut dyn Write) {
     ///         write!(s, "MY_FUNCTION").unwrap();
     ///     }
     /// }