+++ src/query/insert.rs
@@ -557,13 +557,8 @@ impl InsertStatement {
 }
 
 impl QueryStatementBuilder for InsertStatement {
-    fn build_collect_any_into(
-        &self,
-        query_builder: &dyn QueryBuilder,
-        sql: &mut SqlWriter,
-        collector: &mut dyn FnMut(Value),
-    ) {
-        query_builder.prepare_insert_statement(self, sql, collector);
+    fn build_collect_any_into(&self, query_builder: &dyn QueryBuilder, sql: &mut dyn SqlWriter) {
+        query_builder.prepare_insert_statement(self, sql);
     }
 
     fn into_sub_query_statement(self) -> SubQueryStatement {
@@ -572,46 +567,7 @@ impl QueryStatementBuilder for InsertStatement {
 }
 
 impl QueryStatementWriter for InsertStatement {
-    /// Build corresponding SQL statement for certain database backend and collect query parameters
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use sea_query::{tests_cfg::*, *};
-    ///
-    /// let query = Query::insert()
-    ///     .into_table(Glyph::Table)
-    ///     .columns([Glyph::Aspect, Glyph::Image])
-    ///     .values_panic(vec![3.1415.into(), "041080".into()])
-    ///     .to_owned();
-    ///
-    /// assert_eq!(
-    ///     query.to_string(MysqlQueryBuilder),
-    ///     r#"INSERT INTO `glyph` (`aspect`, `image`) VALUES (3.1415, '041080')"#
-    /// );
-    ///
-    /// let mut params = Vec::new();
-    /// let mut collector = |v| params.push(v);
-    ///
-    /// assert_eq!(
-    ///     query.build_collect(MysqlQueryBuilder, &mut collector),
-    ///     r#"INSERT INTO `glyph` (`aspect`, `image`) VALUES (?, ?)"#
-    /// );
-    /// assert_eq!(
-    ///     params,
-    ///     vec![
-    ///         Value::Double(Some(3.1415)),
-    ///         Value::String(Some(Box::new(String::from("041080")))),
-    ///     ]
-    /// );
-    /// ```
-    fn build_collect<T: QueryBuilder>(
-        &self,
-        query_builder: T,
-        collector: &mut dyn FnMut(Value),
-    ) -> String {
-        let mut sql = SqlWriter::new();
-        query_builder.prepare_insert_statement(self, &mut sql, collector);
-        sql.result()
+    fn build_collect_into<T: QueryBuilder>(&self, query_builder: T, sql: &mut dyn SqlWriter) {
+        query_builder.prepare_insert_statement(self, sql);
     }
 }